Extract InsightCard to shared component, add InsightsTabPanel, add hostInsights query key
- Move InsightCard from InsightsPage into components/insights/InsightCard.tsx with
  local expand state (no parent Set plumbing needed).
- Create InsightsTabPanel shared component for reuse across host and container tabs.
- Add queryKeys.hostInsights factory.
- Update InsightsPage to use the new self-contained InsightCard.
